+++ pymodbus/client/asynchronous/mixins.py
@@ -1,13 +1,12 @@
 import logging
-import asyncio
 from pymodbus.client.sync import BaseModbusClient
-from pymodbus.bit_read_message import *
-from pymodbus.bit_write_message import *
-from pymodbus.register_read_message import *
-from pymodbus.register_write_message import *
-from pymodbus.diag_message import *
-from pymodbus.file_message import *
-from pymodbus.other_message import *
+# from pymodbus.bit_read_message import *
+# from pymodbus.bit_write_message import *
+# from pymodbus.register_read_message import *
+# from pymodbus.register_write_message import *
+# from pymodbus.diag_message import *
+# from pymodbus.file_message import *
+# from pymodbus.other_message import *
 from pymodbus.constants import Defaults
 
 from pymodbus.factory import ClientDecoder
@@ -35,16 +34,6 @@ class BaseAsyncModbusClient(BaseModbusClient):
         super(BaseAsyncModbusClient, self).__init__(
             framer or ModbusSocketFramer(ClientDecoder()), **kwargs
         )
-
-    async def execute(self, request=None):
-        """
-        Executes requests asynchronously
-        :param request:
-        :return:
-        """
-        req = self._execute(request)
-        resp = await asyncio.wait_for(req, timeout=self._timeout)
-        return resp
 
 
 class AsyncModbusClientMixin(BaseAsyncModbusClient):
