+++ requirements.txt
@@ -1,4 +1,4 @@
-six>=1.11.0
+six>=1.15.0
 # -------------------------------------------------------------------
 # if want to use the pymodbus serial stack, uncomment these
 # -------------------------------------------------------------------
